[Streams] Partitioning improvements  (#209095)
## Summary

This issue predominantly tries to improve the situation around fetching
and showing samples. Some of the discussion can be seen here:
https://github.com/elastic/streams-program/issues/37#issuecomment-2605288052

We have several issues - runtime fields are expensive (but needed if
fields aren't mapped), we are susceptible to timeouts depending on
amount of data and timerange, getting exact document counts (for match /
not matched counts) is expensive etc.

After speaking with Joe we decided it might be worth trying out async
search, as this alleviates some of these issues. E.g. the ability to
load and show partial results without trying to communicate this through
our API, or have to provide a potentially confusing UI around timeouts /
running to exhaustion options / toggles.

Realistically we only fetch 100 examples, but we might need to scan many
documents to gather that set of documents, I'm not 100% sure how often
we'll actually hit partial results here, but it seems more robust than
worrying about timeouts.

For the matching counts I just couldn't see a way to get an accurate
count without something expensive (e.g. `track_total_hits`) so I've
tried to use an "approximate match rate" based on a random sample, that
random sample is then filtered to the condition to see what approximate
percent matched. One note: aggregations don't seem to return partial
results (which makes sense I guess), you get the interval polling
requests, but won't get a result until the end. I did wonder if you
could do something smart with `track_total_hits` and aggs to "stream"
partial counts, I found a Slack thread saying don't do this 😅

⚠️ ~I'm not 100% sure what I'm missing here but I have seen the filter
sub aggregation come back with a doc_count that is higher than the
random sample.~

~[From the
docs](https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-random-sampler-aggregation.html#random-sampler-inner-workings)
I understand "If a query is provided, a document is returned if it is
matched by the query and if the document is in the random sampling. The
sampling is not done over the matched documents." but I don't see why
that affects the sub aggregation under the random sample.~

![Screenshot 2025-01-31 at 11 30
53](https://github.com/user-attachments/assets/e2444348-caef-41b6-9708-4fdbb84f1ccd)

~I hit this when playing with the `probability` setting, not sure if I'm
missing something stupid.~

[Solved](https://github.com/elastic/kibana/pull/209095#discussion_r1940567855)

Overall, this does seem to work well. I've used this against ~250k and
~2.5million documents, and whilst (depending on time range / runtime
fields) it can still be slow, it seems to provide a better experience
than hitting our API and holding the open connection. Obviously it comes
with the downsides of sitting on the client (not really sure it's a con,
these are platform services) and not using the standard
`streamsRepositoryClient`.

## Other changes

- The core changes here are in the `use_async_sample` hook, and where
that's consumed.

- Runtime fields are not generated for fields that are mapped.

- I've also refactored the routing index page so that components / hooks
live in their own files (this makes the diff look bigger than it is)

- Refactored some logic around preview panel / preview panel
illustration so that the two branches of logic / conditionals now become
one.

## Followups

- I haven't changed enrichment to use this or removed the actual API
route as I figured this would need discussion first to see if we want to
use this.

(cherry picked from commit 97d0c1b2aeee10bdadede71b05691f8857c5fc2f)
