+++ x-pack/solutions/observability/packages/kbn-streams-schema/src/helpers/condition_to_query_dsl.ts
@@ -6,12 +6,12 @@
  */
 
 import {
-  Condition,
   FilterCondition,
-  isAndCondition,
+  Condition,
   isFilterCondition,
+  isAndCondition,
   isOrCondition,
-} from '@kbn/streams-schema';
+} from '../models';
 
 function conditionToClause(condition: FilterCondition) {
   switch (condition.operator) {
